code

+++ paper_rl/logger/mpi_logger.py
@@ -12,6 +12,7 @@ import numpy as np
 import pandas as pd
 import torch
 
+from paper_rl.common.mpi.mpi_tools import mpi_statistics_scalar, proc_id
 from paper_rl.common.mpi.serialization_utils import convert_json
 
 color2num = dict(
@@ -43,7 +44,7 @@ def colorize(string, color, bold=False, highlight=False):
     return "\x1b[%sm%s\x1b[0m" % (";".join(attr), string)
 
 
-class Logger:
+class MPILogger:
     """
     Logging tool
     """
@@ -70,9 +71,10 @@ class Logger:
         self.exp_path = osp.join(workspace, exp_name)
         self.log_path = osp.join(self.exp_path, "logs")
         self.raw_log_file = osp.join(self.log_path, "raw.csv")
-        if clear_out:
-            if osp.exists(self.exp_path):
-                shutil.rmtree(self.exp_path, ignore_errors=True)
+        if proc_id() == 0:
+            if clear_out:
+                if osp.exists(self.exp_path):
+                    shutil.rmtree(self.exp_path, ignore_errors=True)
 
         Path(self.log_path).mkdir(parents=True, exist_ok=True)
 
@@ -95,25 +97,27 @@ class Logger:
         """
         save configuration of experiments to the experiment directory
         """
-        config_path = osp.join(self.exp_path, "config.json")
-        config_json = convert_json(config)
-        output = json.dumps(config_json, indent=2, sort_keys=True)
-        if verbose > 1:
-            self.print("Saving config:\n", color="cyan", bold=True)
-        if verbose > 1:
-            self.print(output)
-        with open(config_path, "w") as out:
-            out.write(output)
+        if proc_id() == 0:
+            config_path = osp.join(self.exp_path, "config.json")
+            config_json = convert_json(config)
+            output = json.dumps(config_json, indent=2, sort_keys=True)
+            if verbose > 1:
+                self.print("Saving config:\n", color="cyan", bold=True)
+            if verbose > 1:
+                self.print(output)
+            with open(config_path, "w") as out:
+                out.write(output)
 
     def print(self, msg, file=sys.stdout, color="", bold=False):
         """
         print to terminal, stdout by default. Ensures only the main process ever prints.
         """
-        if color == "":
-            print(msg, file=file)
-        else:
-            print(colorize(msg, color, bold=bold), file=file)
-        sys.stdout.flush()
+        if proc_id() == 0:
+            if color == "":
+                print(msg, file=file)
+            else:
+                print(colorize(msg, color, bold=bold), file=file)
+            sys.stdout.flush()
 
     def store(self, tag="default", append=True, **kwargs):
         """
@@ -146,19 +150,20 @@ class Logger:
         return self.data[tag]
 
     def pretty_print_table(self, data):
-        vals = []
-        key_lens = [len(key) for key in data.keys()]
-        max_key_len = max(15, max(key_lens))
-        keystr = "%" + "%d" % max_key_len
-        fmt = "| " + keystr + "s | %15s |"
-        n_slashes = 22 + max_key_len
-        print("-" * n_slashes)
-        for key in data.keys():
-            val = data[key]
-            valstr = "%8.3g" % val if hasattr(val, "__float__") else val
-            print(fmt % (key, valstr))
-            vals.append(val)
-        print("-" * n_slashes, flush=True)
+        if proc_id() == 0:
+            vals = []
+            key_lens = [len(key) for key in data.keys()]
+            max_key_len = max(15, max(key_lens))
+            keystr = "%" + "%d" % max_key_len
+            fmt = "| " + keystr + "s | %15s |"
+            n_slashes = 22 + max_key_len
+            print("-" * n_slashes)
+            for key in data.keys():
+                val = data[key]
+                valstr = "%8.3g" % val if hasattr(val, "__float__") else val
+                print(fmt % (key, valstr))
+                vals.append(val)
+            print("-" * n_slashes, flush=True)
 
     def log(self, step):
         """
@@ -170,26 +175,27 @@ class Logger:
         for tag in self.data.keys():
             data_dict = self.data[tag]
             for k, v in data_dict.items():
-                if isinstance(v, list):
-                    vals = np.array(v)
-                    vals_sum, n = vals.sum(), len(vals)
-                    avg = vals_sum / n
-                    sum_sq = np.sum((vals - avg) ** 2)
-                    std = np.sqrt(sum_sq / n)
-                    minv = np.min(vals)
-                    maxv = np.max(vals)
+                if isinstance(v, int) or isinstance(v, float):
+                    key_vals = {f"{tag}/{k}": v}
+                else:
+                    vals = (
+                        np.concatenate(v)
+                        if isinstance(v[0], np.ndarray) and len(v[0].shape) > 0
+                        else v
+                    )
+                    stats = mpi_statistics_scalar(vals, with_min_and_max=True)
+                    avg, std, minv, maxv = stats[0], stats[1], stats[2], stats[3]
                     key_vals = {
                         f"{tag}/{k}_avg": avg,
                         f"{tag}/{k}_std": std,
                         f"{tag}/{k}_min": minv,
                         f"{tag}/{k}_max": maxv,
                     }
-                else:
-                    key_vals = {f"{tag}/{k}": v}
-                for name, scalar in key_vals.items():
-                    if self.tensorboard:
-                        self.tb_writer.add_scalar(name, scalar, step)
-                    self.stats[name] = scalar
+                if proc_id() == 0:
+                    for name, scalar in key_vals.items():
+                        if self.tensorboard:
+                            self.tb_writer.add_scalar(name, scalar, step)
+                        self.stats[name] = scalar
         return self.stats
 
     def reset(self):
